+++ app/src/main/java/com/freshdigitable/yttt/SubscriptionListFragment.kt
@@ -4,10 +4,6 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.itemsIndexed
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.Fragment
@@ -16,10 +12,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.asLiveData
 import androidx.lifecycle.viewModelScope
-import androidx.navigation.NavController
-import androidx.navigation.compose.rememberNavController
 import androidx.navigation.findNavController
-import com.freshdigitable.yttt.compose.LiveChannelListItemView
+import com.freshdigitable.yttt.compose.SubscriptionListScreen
 import com.freshdigitable.yttt.data.YouTubeLiveRepository
 import com.freshdigitable.yttt.data.model.LiveSubscription
 import com.google.accompanist.themeadapter.material.MdcTheme
@@ -55,25 +49,4 @@ class SubscriptionListViewModel @Inject constructor(
 ) : ViewModel() {
     val subscriptions: LiveData<List<LiveSubscription>> =
         repository.subscriptions.asLiveData(viewModelScope.coroutineContext)
-}
-
-@Composable
-fun SubscriptionListScreen(
-    viewModel: SubscriptionListViewModel,
-    navController: NavController = rememberNavController()
-) {
-    val subs = viewModel.subscriptions.observeAsState().value ?: emptyList()
-    LazyColumn {
-        itemsIndexed(items = subs, key = { _, item -> item.id.value }) { _, item ->
-            LiveChannelListItemView(
-                iconUrl = item.channel.iconUrl,
-                title = item.channel.title,
-                onClick = {
-                    val action = SubscriptionListFragmentDirections
-                        .actionMenuSubscriptionListToChannelFragment(item.channel.id.value)
-                    navController.navigate(action)
-                },
-            )
-        }
-    }
 }
